requirements.txt.txt
django
djangorestframework
psycopg2
playwright
SQLAlchemy
selenium
chromedriver-autoinstaller
beautifulsoup4
pandas
requests